requirements.txt
django_heroku==0.3.1
Django==2.1.4
djangorestframework==3.11.0
gunicorn==20.0.4